Make search input take up the whole container
Previously, the form and, consequently, the input in the browser page
doesn't take the whole width of its parent container, so when user
types a very long URL, most of the text will not be visible--even if
the container is big enough. To fix this, I added width:100%; to the
form.

Additionally, the searchInput class in Home.module.css is used for both
the input container and the form, which is not a good idea. So, I've
seperated the styles of the form into another ruleset.